moved utilities functions
